video_sorting_1_0
video_sorting_1_0 - program for sorting video by download date and renaming each video file to a new name in the format of "01.mp4", "02.mp4", etc.

This program asks the user to input the directory path where the video files are located. It then loops through all the files in the directory and adds the video files to a list along with their modified times. The list is sorted based on the modified times of the files. It then loops through the sorted video files list and renames each video file to a new name in the format of "01.mp4", "02.mp4", and so on. The user can choose any directory containing video files to run this program.